[STEP 07/08] 이커머스 비즈니스 로직 개발 및 테스트 작성

+++ src/main/java/kr/hhplus/be/commerce/interfaces/dto/requestDto/OrderItemRequest.java
@@ -5,6 +5,6 @@ import lombok.Builder;
 @Builder
 public record OrderItemRequest(
         Long productId,
-        Integer quantity
+        Long quantity
 ) {
 }
